rename(HomeFragment): 클래스명 수정
- before: HomeFragment
- after: ScreeningMovieFragment
