+++ app/src/main/java/woowacourse/movie/presentation/view/navigation/ScreeningMovieFragment.kt
@@ -15,7 +15,7 @@ import woowacourse.movie.presentation.view.screening.ScreeningPresenterImpl
 import woowacourse.movie.presentation.view.screening.adapter.MovieListAdapter
 import woowacourse.movie.presentation.view.screening.theater.TheaterBottomSheetDialogFragment
 
-class HomeFragment : Fragment(), ScreeningContract.View, ScreeningContract.ViewActions {
+class ScreeningMovieFragment : Fragment(), ScreeningContract.View, ScreeningContract.ViewActions {
     private lateinit var adapter: MovieListAdapter
     private lateinit var presenter: ScreeningContract.Presenter
